requirements.txt
flask
pandas
scikit-learn
pytesseract
pillow
pymongo
matplotlib
seaborn
